+++ connecteur/test/action/PastellCoreTestFatalError.php
@@ -1,9 +1,14 @@
 <?php
 
+declare(strict_types=1);
+
 class PastellCoreTestFatalError extends ActionExecutor
 {
-    public function go()
+    /**
+     * @throws UnrecoverableException
+     */
+    public function go(): void
     {
-        trigger_error("Fatal error", E_USER_ERROR);
+        throw new UnrecoverableException('Fatal error');
     }
 }